- Add Purchase Reducer - Add Actions to Load - Warning fix
