+++ peddle-webapp/src/actions/actionsTypes.js
@@ -5,9 +5,8 @@ export const LOGGED_IN = 'LOGGED_IN';
 export const LOAD_WISHLIST = 'LOAD_WISHLIST';
 export const ADD_TO_WISHLIST = 'ADD_TO_WISHLIST';
 export const REMOVE_FROM_WISHLIST = 'REMOVE_FROM_WISHLIST';
-export const LOAD_PURCHACE_LIST = 'LOAD_PURCHACE_LIST';
-export const ADD_TO_PURCHACE_LIST = 'ADD_TO_PURCHACE_LIST';
-export const REMOVE_FROM_PURCHACE_LIST = 'REMOVE_FROM_PURCHACE_LIST';
+export const LOAD_PURCHASE_LIST = 'LOAD_PURCHASE_LIST';
+
