+++ peddle-webapp/src/reducers/purchaseReducer.js
@@ -1,13 +1,9 @@
-import { ADD_TO_PURCHACE_LIST,LOAD_PURCHACE_LIST, REMOVE_FROM_PURCHACE_LIST} from "../actions/actionsTypes";
+import { LOAD_PURCHASE_LIST } from "../actions/actionsTypes";
 
 function purchaseReducer(state=[], action) {
     switch (action.type) {
-        case LOAD_PURCHACE_LIST:
+        case LOAD_PURCHASE_LIST:
             return Object.assign({}, state, action.data);
-        case ADD_TO_PURCHACE_LIST:
-            return state;
-        case REMOVE_FROM_PURCHACE_LIST:
-            return state;
         default:
             return state;
     }
